+++ database/migrations/2020_04_30_052826_create_order_details_table.php
@@ -33,7 +33,7 @@ class CreateOrderDetailsTable extends Migration
 
 //            $table->integer('job_status')->nullable('false')->default(0);
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table ->foreign('status_id')->references('id')->on('status_types');
             $table->timestamps();
         });
